Extract components into their own files
